requirements.txt
streamlit==1.19.0
langcodes==3.3.0
SpeechRecognition==3.9.0
json5==0.9.6
jsonschema==3.2.0
pydub==0.25.1
google-cloud==0.34.0
google-cloud-speech==1.1.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-api-core==1.34.0
google-api-python-client==1.7.11